This update increases the move the buffer size macro from the main .h to the platform specific .h files, allowing tweaking of that parameter on a per platform basis. This now allows us to use a more generous buffer for the Arduino Due that would not be advisable on older platforms.
This update also introduces a new class: XbeeWifiBuffered.

This class derives from the main XbeeWifi class and can be used to achieve a simplified read pattern based on the Arduino standards of available() and read(), plus associated methods.

This does use more memory, however, on the Arduino Due and larger ATMEGA based platforms this would be acceptible. The extra memory use is to buffer the incoming data to allow it's use with the simplified methods.